move useViewScale to helpers.ts
