+++ web/pages/week/[username]/[rangeEndDateSlug].tsx
@@ -15,7 +15,7 @@ import { query, where } from 'firebase/firestore'
 import { chunk, orderBy, sortBy, sum } from 'lodash'
 import { useMemo } from 'react'
 import { CopyLinkOrShareButton } from 'web/components/buttons/copy-link-button'
-import { useViewScale } from 'web/components/charts/generic-charts'
+import { useViewScale } from 'web/components/charts/helpers'
 import { ContractsGrid } from 'web/components/contract/contracts-grid'
 import { ProfitChangeTable } from 'web/components/daily-profit'
 import { Col } from 'web/components/layout/col'
